+++ mercicor/processing/calcul/calcul_pertes_gains.py
@@ -41,11 +41,13 @@ class BaseCalculPertesGains(CalculAlgorithm):
     def name(self):
         return 'calcul_{}'.format(self.project_type.label)
 
+    # noinspection PyPep8Naming
     def displayName(self):
         return (
             'Calcul des notes de {} pour le scénario de {}'.format(
                 self.project_type.calcul_type, self.project_type.label))
 
+    # noinspection PyPep8Naming
     def checkParameterValues(self, parameters, context):
         sources = [
             self.parameterAsVectorLayer(parameters, self.SCENARIO_IMPACT, context),
@@ -81,6 +83,7 @@ class BaseCalculPertesGains(CalculAlgorithm):
             )
         return message
 
+    # noinspection PyMethodOverriding
     def initAlgorithm(self, config):
         _ = config
 
@@ -118,24 +121,38 @@ class BaseCalculPertesGains(CalculAlgorithm):
                 field_name = '{calcul_type}_{note}'.format(
                     calcul_type=self.project_type.calcul_type, note=note)
                 feat[field_name] = 0
-                feedback.pushDebugInfo("Initialisation scénario {} {} = 0".format(scenario_id, field_name))
+                feedback.pushDebugInfo(
+                    "Initialisation scénario {} champ {} = 0".format(scenario_id, field_name))
 
                 filter_expression = QgsExpression.createFieldEqualityExpression('scenario_id', scenario_id)
                 for feature in hab_etat_ecolo.getFeatures(filter_expression):
 
+                    feedback.pushInfo(
+                        "Lecture d'une nouvelle entité habitat état écologique ID {} dont le "
+                        "scénario = {}".format(
+                            hab_etat_ecolo.id(), scenario_id))
+
                     if feature[fields[1]] == NULL:
                         feedback.pushDebugInfo(
-                            "Omission du calcul {} pour l'entité {}".format(field_name, feature.id()))
+                            "Omission du calcul {} pour l'entité {} car le champ est NULL".format(
+                                field_name, feature.id()))
                         continue
 
                     if self.project_type == ProjectType.Pression:
                         sub_result = feature[fields[0]] - feature[fields[1]]
+                        feedback.pushDebugInfo("Sous résultat {} - {} = {}".format(
+                            fields[0], fields[1], sub_result))
                         divide = 1
                     else:
                         # Compensation
-                        # Tenir compte du délais et du risque
+                        # Tenir compte des délais et du risque
                         sub_result = feature[fields[1]] - feature[fields[0]]
+                        feedback.pushDebugInfo(
+                            "Sous résultat {} - {} = {}".format(fields[1], fields[0], sub_result))
                         divide = feature['compensation_coeff_risque'] * feature['compensation_coeff_delais']
+                        feedback.pushDebugInfo(
+                            "Division compensation_coeff_risque * compensation_coeff_delais = {}".format(
+                                divide))
 
                     if len(fields) == 3:
                         # Pour le calcul de perte_bsd, perte_bsm, gain_bsd et gain_bsm,
@@ -143,29 +160,37 @@ class BaseCalculPertesGains(CalculAlgorithm):
                         multiply = feature[fields[2]]
                     else:
                         multiply = 1
+                    feedback.pushDebugInfo("Coefficient multiplicateur : {}".format(multiply))
 
-                    feat[field_name] += multiply * (sub_result * feature.geometry().area()) / divide
+                    temporary_result = multiply * (sub_result * feature.geometry().area()) / divide
+                    feat[field_name] += temporary_result
 
                     # Tentative explication de la formule
                     feedback.pushDebugInfo(
                         "Scénario {scenario_id} entité habitat état écologique {id} : "
-                        "{field_name} += {multiply} * ( {sub_result} * surface) / {divide}".format(
+                        "{field_name} += {multiply} * ( {sub_result} * surface {surface}) / {divide}".format(
                             scenario_id=scenario_id,
                             id=feature.id(),
                             field_name=field_name,
                             multiply=multiply,
                             sub_result=sub_result,
+                            surface=feature.geometry().area(),
                             divide=divide,
                         )
                     )
+                    feedback.pushDebugInfo(
+                        "Incrémentation scénario ID {} champ {} de {}. Nouvelle valeur temporaire {}. "
+                        "Passage à l'entité suivante…".format(
+                            scenario_id, field_name, temporary_result, feat[field_name]))
 
                 feedback.pushDebugInfo(
-                    "Fin sur le scénario ID {} champ {} = {}".format(
-                        scenario_id, field_name, feat[field_name]))
+                    "Fin sur le scénario ID {} champ {} = {}, passage au champ suivant avec le même "
+                    "scénario…".format(scenario_id, field_name, feat[field_name]))
 
             scenario_impact.updateFeature(feat)
-            feedback.pushDebugInfo("Fin sur le scénario ID {}".format(scenario_id))
-        feedback.pushInfo("Fin des calculs, enregistrement…")
+            feedback.pushDebugInfo("Fin sur le scénario ID {}, passage au scénario suivant…".format(
+                scenario_id))
+        feedback.pushInfo("Fin des calculs, enregistrement de la couche…")
         scenario_impact.commitChanges()
 
         return {}